Move NavigatorDefaults to NavigableCircuitContent for better dep splitting
