+++ circuit/build.gradle.kts
@@ -39,6 +39,8 @@ kotlin {
     maybeCreate("androidMain").apply {
       dependencies {
         api(libs.androidx.compose.runtime)
+        api(libs.androidx.compose.animation)
+        api(libs.androidx.compose.foundation)
         implementation(libs.androidx.compose.integration.activity)
       }
     }
