+++ circuit/src/androidMain/kotlin/com/slack/circuit/NavigableCircuitContent.android.kt
@@ -16,17 +16,31 @@
 package com.slack.circuit
 
 import androidx.activity.compose.BackHandler
+import androidx.compose.animation.AnimatedContent
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.animation.SizeTransform
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.slideInHorizontally
+import androidx.compose.animation.slideOutHorizontally
+import androidx.compose.animation.with
+import androidx.compose.foundation.background
+import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.key
 import androidx.compose.runtime.movableContentOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
 import com.slack.circuit.backstack.BackStack
 import com.slack.circuit.backstack.NavDecoration
-import com.slack.circuit.backstack.NavigatorDefaults
 import com.slack.circuit.backstack.ProvidedValues
 import com.slack.circuit.backstack.SaveableBackStack
 import com.slack.circuit.backstack.isAtRoot
@@ -93,5 +107,80 @@ public fun BasicNavigableCircuitContent(
       val providedLocals = remember(values) { values?.toTypedArray() ?: emptyArray() }
       CompositionLocalProvider(*providedLocals) { provider() }
     }
+  }
+}
+
+/** Default values and common alternatives used by navigable composables. */
+public object NavigatorDefaults {
+
+  private const val FIVE_PERCENT = 0.05f
+  private val SlightlyRight = { width: Int -> (width * FIVE_PERCENT).toInt() }
+  private val SlightlyLeft = { width: Int -> 0 - (width * FIVE_PERCENT).toInt() }
+
+  /** The default [NavDecoration] used in navigation. */
+  public object DefaultDecoration : NavDecoration {
+    @OptIn(ExperimentalAnimationApi::class)
+    @Composable
+    override fun <T> DecoratedContent(
+      arg: T,
+      backStackDepth: Int,
+      modifier: Modifier,
+      content: @Composable (T) -> Unit
+    ) {
+      // Remember the previous stack depth so we know if the navigation is going "back".
+      val prevStackDepth = rememberSaveable { mutableStateOf(backStackDepth) }
+      val diff = backStackDepth - prevStackDepth.value
+      prevStackDepth.value = backStackDepth
+      AnimatedContent(
+        targetState = arg,
+        modifier = modifier,
+        transitionSpec = {
+          // Mirror the forward and backward transitions of activities in Android 33
+          if (diff > 0) {
+            slideInHorizontally(tween(), SlightlyRight) + fadeIn() with
+              slideOutHorizontally(tween(), SlightlyLeft) + fadeOut()
+          } else
+            if (diff < 0) {
+                slideInHorizontally(tween(), SlightlyLeft) + fadeIn() with
+                  slideOutHorizontally(tween(), SlightlyRight) + fadeOut()
+              } else {
+                // Crossfade if there was no diff
+                fadeIn() with fadeOut()
+              }
+              .using(
+                // Disable clipping since the faded slide-in/out should
+                // be displayed out of bounds.
+                SizeTransform(clip = false)
+              )
+        },
+      ) {
+        content(it)
+      }
+    }
+  }
+
+  /** An empty [NavDecoration] that emits the content with no surrounding decoration or logic. */
+  public object EmptyDecoration : NavDecoration {
+    @Composable
+    override fun <T> DecoratedContent(
+      arg: T,
+      backStackDepth: Int,
+      modifier: Modifier,
+      content: @Composable (T) -> Unit
+    ) {
+      content(arg)
+    }
+  }
+
+  /**
+   * Bright ugly error text telling a developer they didn't provide a route that a [BackStack] asked
+   * for.
+   */
+  public val UnavailableRoute: @Composable (String) -> Unit = { route ->
+    BasicText(
+      "Route not available: $route",
+      Modifier.background(Color.Red),
+      style = TextStyle(color = Color.Yellow)
+    )
   }
 }
